refs #1 app image 수정
