+++ shake-app/android/app/src/main/kotlin/com/mermer/shake_app/MainActivity.kt
@@ -2,5 +2,4 @@ package com.mermer.shake_app
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
